tests/wpt/web-platform-tests/offscreen-canvas/path-objects/2d.path.arc.negative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.path.arc.negative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.path.arc.negative</h1>
<p class="desc">arc() with negative radius throws INDEX_SIZE_ERR</p>


<script>
var t = async_test("arc() with negative radius throws INDEX_SIZE_ERR");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

assert_throws("INDEX_SIZE_ERR", function() { ctx.arc(0, 0, -1, 0, 0, true); });

t.done();

});
</script>
